+++ src/app/[lang]/@blogs/BlogElement.tsx
